Move `trait Tree` to its own module Also, set the number of children reported by Node::Branch to 16 instead of the number of non-empty nodes.
